Waveshare Modbus Relay Driver (#2085)
* Preliminary driver

* First relay should start from 1

* Driver is operational

* Add driver to list

+++ libs/indibase/defaultdevice.cpp
@@ -838,13 +838,13 @@ bool DefaultDevice::updateProperties()
 // The situation occurs in the case of drivers where the class that inherits from `DefaultDevice`
 // will call the `setDriverInterface` function in the constructor
 
-uint16_t DefaultDevice::getDriverInterface() const
+uint32_t DefaultDevice::getDriverInterface() const
 {
     D_PTR(const DefaultDevice);
     return atoi(d->DriverInfoTP[3 /* DRIVER_INTERFACE */].getText());
 }
 
-void DefaultDevice::setDriverInterface(uint16_t value)
+void DefaultDevice::setDriverInterface(uint32_t value)
 {
     D_PTR(DefaultDevice);
     d->DriverInfoTP[3 /* DRIVER_INTERFACE */].setText(std::to_string(value));
